[#1] Suppress MySQL connection errors
Suppress "Can't connect to local MySQL server through socket" errors.

Connecting to the MySQL server over TCP/IP instead of Unix sockets
doesn't appear to help.

It's not particularly important this section is tested so for now we
will add a testing variable and not include these states if testing is
enabled.

Fixes #3.